fix avatar

+++ src/com/buldreinfo/jersey/jaxb/io/ImageHelper.java
@@ -39,19 +39,25 @@ public class ImageHelper {
 		dao.setMediaMetadata(c, idMedia, imageReader.getJpgBufferedImage().getHeight(), imageReader.getJpgBufferedImage().getWidth(), exifReader.getDateTaken());
 	}
 	
-	public static void saveAvatar(int userId, String avatarUrl) throws IOException, InterruptedException {
+	public static void saveAvatar(int userId, String avatarUrl, boolean replaceOriginal) throws IOException, InterruptedException {
 		Path original = IOHelper.getPathOriginalUsers(userId);
-		IOHelper.createDirectories(original.getParent());
-		try (InputStream in = URI.create(avatarUrl).toURL().openStream()) {
-			Files.copy(in, original, StandardCopyOption.REPLACE_EXISTING);
-			IOHelper.setFilePermission(original);
+		boolean createResized = false;
+		if (replaceOriginal || !Files.exists(original)) {
+			IOHelper.createDirectories(original.getParent());
+			try (InputStream in = URI.create(avatarUrl).toURL().openStream()) {
+				Files.copy(in, original, StandardCopyOption.REPLACE_EXISTING);
+				IOHelper.setFilePermission(original);
+				createResized = true;
+			}
 		}
 		Path resized = IOHelper.getPathWebUsers(userId);
-		IOHelper.deleteIfExistsCreateParent(resized.getParent());
-		BufferedImage b = ImageReader.newBuilder().withPath(original).build().getJpgBufferedImage();
-		b = Scalr.resize(b, Scalr.Method.ULTRA_QUALITY, Scalr.Mode.FIT_EXACT, 35, 35, Scalr.OP_ANTIALIAS);
-		Preconditions.checkArgument(ImageIO.write(b, "jpg", resized.toFile()));
-		IOHelper.setFilePermission(resized);
+		if (createResized || !Files.exists(resized)) {
+			IOHelper.deleteIfExistsCreateParent(resized);
+			BufferedImage b = ImageReader.newBuilder().withPath(original).build().getJpgBufferedImage();
+			b = Scalr.resize(b, Scalr.Method.ULTRA_QUALITY, Scalr.Mode.FIT_EXACT, 35, 35, Scalr.OP_ANTIALIAS);
+			Preconditions.checkArgument(ImageIO.write(b, "jpg", resized.toFile()));
+			IOHelper.setFilePermission(resized);
+		}
 	}
 
 	public static void saveImage(Dao dao, Connection c, int idMedia, BufferedImage bufferedImage) throws IOException, SQLException {
